+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.11.0-exp.1] - 2023-05-26
 
 ### Added
 
+- Added new option to account for solid occluders in wind propagation.
+- Added new option to account for solid occluders in strand count probe construction.
 - Added new component 'HairWind' for authoring wind emitters that can affect hair. The HairWind component represents a single emitter, and can be configured either as a standalone emitter (preferred), or as a proxy for an existing WindZone (more limited).
-- Added support for wind forces via volumetric wind propagation. Wind propagation is enabled per HairInstance under volume settings. When enabled, wind will begin to propagate through hair, and act on hair according to how much wind was able to pass through the volume between a cell and any emitters, taking into account the physical density of hair in the volume.
+- Added support for volumetric wind propagation. Wind propagation is enabled per HairInstance (under volume settings). When enabled, wind will propagate through the volume of hair, and act on strands accordingly, taking into account the physical density of hair in the volume.
 
 ### Fixed